re-align torus-utils-swift
Re-align the current implementation of torus-utils-swift to that of the web implementation of torus.js.

This will reduce the overhead of differences a developer would need to remember between the two and provide an overall better experience.

This will also provide better long term maintainability and make feature parity simpler to implement between the two.

The are also numerous bug fixes implemented in this PR as well, one of which allowed the package AnyCodable as a dependency to be removed.

 Should be feature-complete up until tag v12.3.5 on torus.js